Add files via upload
Does the impressions this visualisation vehiculates seem legitimate? Or are their impressions that don't relate to the data but come from artifices from the conceptor of the visualisation? I say what I think a few lines below.




Each circle is supposed to represent the proportion of Death/Money in that circle compared to the total. However, that proportion is directly proportionnal to the radius of the circle rather then its area. Also, the fact that the biggest circle is associated with the darkest colour is unlikely attribuable to chance. Those are mechanisms  using graphic forms in inappropriate ways.

In principle, we associate the quantity in a circle to be proportionnal to its area. For exemple, if we were in 3D, we would associate the amount of deaths in a cube to be proportionnal to its volume and not to the surface of one of its faces (2D), and even less to the length of one of its sizes (1D). The fact that it is directly proportionnal is convenient: it is a reference. If it would be fare to change this proportionnal relation according to the distribution of the data (e.g. logarythmic for very large data) it should be done explicitly/ with litteral text, and not through geometrical shapes. Also, the legitimacy of this graph depends on the fact that the sum of the circles' area in each genera (Death and Money) is the same, otherwise comparison between each genera is not possible. This supports the convention that the area should represent the quantity rather then the radius.

T​his is misleading because we see the proportions to be the square ofwhat they are: for a circle B containing twice the amount of A, it will seem 4 times bigger then A.  The main message I get from this visualisation is that the amounts of money invested for certain deseases are disproportionnate to the amount of deaths those deseases cause. This disproportion is greatly exagerated.

A​s for the colour, I feel we give more attention to items contrasting with their background, in this case dark purple agains the white of the page. The fact the colours in our visualization are only tokens makes their power of attraction misleading. Generally, the efficiency of colours for tokenizing balances their misleading potential. However, the fact that they accentuate in our visualisation the main element brings us to suspect they were used unduly for their power of attraction. 

I would like to say that this kind of graphic works in very similar way then a Pie Chart, however it makes comparison between specific parts of the chart more precise as you don't get the biases due to the different orientations of the parts in a pie chart. The general interpretation is also more intuitive then the pie chart. However, the interpretation relies on the hypothesis that proportions are respected between the circles' size and the quantities evaluated, and, if comparing charts, that the sum of the circles' area in each chart has to be the same. Both of these hypothesis are much harder to prove by eye then their corollary in a pie chart (e.g. the pie is full and 36% covers about a third of the chart)